+++ src/theia/TheiaUtils.sol
@@ -16,18 +16,12 @@ library TheiaUtils {
         return 0;
     }
 
-    function hexStringToAddress(
-        string memory s
-    ) internal pure returns (bytes memory) {
+    function hexStringToAddress(string memory s) internal pure returns (bytes memory) {
         bytes memory ss = bytes(s);
         require(ss.length % 2 == 0); // length must be even
         bytes memory r = new bytes(ss.length / 2);
-        for (uint i = 0; i < ss.length / 2; ++i) {
-            r[i] = bytes1(
-                fromHexChar(uint8(ss[2 * i])) *
-                    16 +
-                    fromHexChar(uint8(ss[2 * i + 1]))
-            );
+        for (uint256 i = 0; i < ss.length / 2; ++i) {
+            r[i] = bytes1(fromHexChar(uint8(ss[2 * i])) * 16 + fromHexChar(uint8(ss[2 * i + 1])));
         }
 
         return r;
@@ -39,10 +33,7 @@ library TheiaUtils {
         address tempAddress;
 
         assembly {
-            tempAddress := div(
-                mload(add(add(_bytes, 0x20), 1)),
-                0x1000000000000000000000000
-            )
+            tempAddress := div(mload(add(add(_bytes, 0x20), 1)), 0x1000000000000000000000000)
         }
 
         return tempAddress;
